Jan_22_2019

+++ StringPullingMouse_2_1/Functions/masks/getThisMask.m
@@ -2,7 +2,7 @@ function Is = getThisMask(hsvFrame,theseColors,nrows,ncols,NNs)
 if ~exist('NNs','var')
     NNs = 500;
 end
-lFrame = reshape(hsvFrame,nrows*ncols,3);
+lFrame = double(reshape(hsvFrame,nrows*ncols,3));
 [Idx,D] = knnsearch(lFrame,theseColors,'K',NNs);
 % [Idx,D] = knnsearch(lFrame,uTC,'K',100);
 tZ = zeros(size(lFrame,1),1);
